src/telegram_bot/requirements.txt
gunicorn
aiogram==2.20
aioredis==1.3.1
#redis>=4.2.0rc1
aiohttp_cors
aiohttp_jinja2
jinja2
#aiohttp-devtools
anyio
attrs
chardet
dadata
environs
h11
hiredis
httpcore
httpx
humanfriendly
idna
marshmallow
multidict
parse
phonenumberslite
pydantic==1.8.1
python-dateutil
python-dotenv
python-stdnum
pytz
requests
rfc3986
six
sniffio
typing-extensions
urllib3
yarl
aiohttp-devtools
